chore: disable download expiration

+++ apps/tasks/src/jobs/cleanDownloads.ts
@@ -6,7 +6,7 @@ import { TaskJob } from '../types';
 
 const downloadConfig = config.get('downloads') as {
   path: string;
-  expiration: number;
+  expiration?: number;
 };
 
 export default class CleanDownloads extends TaskJob {
@@ -18,14 +18,16 @@ export default class CleanDownloads extends TaskJob {
     this.logger.log('Cleaning downloads...');
     const downloadPath = path.join(__dirname, '..', '..', downloadConfig.path);
 
-    for (const file of await readdir(downloadPath)) {
-      try {
-        const s = await stat(path.join(downloadPath, file));
-        if (s.mtime.getTime() + downloadConfig.expiration < Date.now()) {
-          this.logger.log(`Deleting ${file}`);
-          await unlink(path.join(downloadPath, file));
-        }
-      } catch (e) {}
+    if (downloadConfig.expiration && downloadConfig.expiration > 0) {
+      for (const file of await readdir(downloadPath)) {
+        try {
+          const s = await stat(path.join(downloadPath, file));
+          if (s.mtime.getTime() + downloadConfig.expiration < Date.now()) {
+            this.logger.log(`Deleting ${file}`);
+            await unlink(path.join(downloadPath, file));
+          }
+        } catch (e) {}
+      }
     }
     this.logger.info('OK.');
   }
